Unwind with pointer authentication on arm64
This patch adds the ability for libunwind to unwind a stack where the
return address obtained from the arm64 link register (x30) has
a pointer authentication code (PAC).

Without this patch, the unw_step function terminates early, leaving the
user the impression there is only a leaf frame when there are more.

The reason for the premature termination is that an aarch64 specific
CFI opcode 'DW_CFA_AARCH64_negate_ra_state' is not recognised, and
treated as unexpected.  Upon the next iteration's call of unw_step, it
reports the end of the stack.

For an expected callstack of

  loop   // most recent call
  call3
  call2
  call1
  main   // oldest call

The caller of libunwind observes:

  unw_step() = 1
  unw_get_reg(UNW_REG_IP) = 0, pc = 0xaaaae4720798
  unw_step() = 1
  unw_get_reg(UNW_REG_IP) = 0, pc = 0x5faaaae47207cc // ip with PAC
  unw_step() = 0  // premature end

For remote unwinding, this patch adds an optional callback function
'ptrauth_insn_mask' on unw_accessors_t so that unw_step can correctly
strip off the PAC to leave the correct return address.  This also has
the benefit that an attempt to get the name of the function with
unw_get_proc_name can succeed.

The callback is given the original unw_addr_space_t and as_arg as given
to unw_init_remote.  The application needs to ensure the mask to be
returned is somehow available on the as_arg and callback should return
it.

The callback is only needed for remote unwinding.  In local unwinding,
an instruction is used to strip the PAC.

A description of pointer authentication is in the armv8 reference manual
available here:

  https://developer.arm.com/documentation/ddi0487/ha/?lang=en

(Version H.a, section D5.1.5 Pointer authentication in AArch64 state,
page D5-4772)

Also documentation for the use of DWARF with pointer authentication, in
particular the RA_SIGN_STATE:

  https://github.com/ARM-software/abi-aa/blob/main/aadwarf64/aadwarf64.rst

(cherry picked from commit f67ef2867bc5e74cdadb3acb790c3fc6161b22e9)